+++ src/components/Checkbox/index.js
@@ -1,16 +1,15 @@
-/*
- * Checkbox
- * Custom checkbox element
- */
+//
+// Checkbox
+// Custom checkbox element
+//
 
-import React from 'react'
-import PropTypes from 'prop-types'
-import classNames from 'classnames'
+import React from "react"
+import PropTypes from "prop-types"
+import cx from "classnames"
+import Label from "components/Label"
 
-import Label from 'components/Label'
-
-
-const Checkbox = (props) => {
+// component
+const Checkbox = props => {
 
   const {
     checked,
@@ -21,10 +20,10 @@ const Checkbox = (props) => {
     id,
     labelProps,
     required,
-    ...other
+    ...attributes
   } = props
 
-  const classes = classNames (
+  const classes = cx (
     "custom-control custom-checkbox",
     className
   )
@@ -38,7 +37,7 @@ const Checkbox = (props) => {
         checked={checked}
         disabled={disabled}
         required={required}
-        {...other}/>
+        {...attributes}/>
       <Label className="custom-control-label" hidden={hideLabel} htmlFor={id} {...labelProps}>
         {children}
       </Label>
@@ -53,6 +52,10 @@ Checkbox.propTypes = {
   id: PropTypes.string,
   labelProps: PropTypes.object,
   required: PropTypes.bool
+}
+
+Checkbox.defaultProps = {
+  id: "checkbox"
 }
 
 export default Checkbox