+++ src/components/Link/index.js
@@ -1,47 +1,43 @@
-/*
- * Link
- * Render common hyperlinks
- */
+//
+// Link
+// Render common hyperlinks
+//
 
-import React from 'react'
-import PropTypes from 'prop-types'
-import classNames from 'classnames'
+import React from "react"
+import PropTypes from "prop-types"
+import cx from "classnames"
+import { CommonLinks } from "utils"
 
-// Core
-import { CommonLinks } from 'utils'
-
-
-const Link = (props) => {
+// component
+const Link = props => {
 
   const {
     blank,
     className,
     children,
     href,
-    id,
     onClick,
     type,
-    ...other
+    ...attributes
   } = props
 
-  const classes = classNames(
+  const classes = cx(
     (!href && !type) && "cursor-pointer",
     className
   )
 
-  const link = CommonLinks[type],
-        target = blank ? "_blank" : "_self"
+  const link = CommonLinks[type]
+  const target = blank ? "_blank" : "_self"
 
   return (
     <a
-      id={id}
       className={classes}
       href={link ? link.href : href}
       rel={blank ? "noopener noreferrer" : "nofollow"}
       onClick={onClick}
       tabIndex="-1"
       target={href && target}
-      {...other}>
+      {...attributes}>
       {children || link && link.title}
     </a>
   )
@@ -50,7 +46,6 @@ const Link = (props) => {
 Link.propTypes = {
   blank: PropTypes.bool,
   href: PropTypes.string,
-  id: PropTypes.string,
   onClick: PropTypes.func,
   type: PropTypes.string
 }