+++ src/components/Logo/index.js
@@ -1,26 +1,26 @@
-/*
- * Logo
- * Render brand logo svgs
- */
+//
+// Logo
+// Render brand logo svgs
+//
 
-import React from 'react'
-import PropTypes from 'prop-types'
-import classNames from 'classnames'
+import React from "react"
+import PropTypes from "prop-types"
+import cx from "classnames"
 
-// Assets
-import LogoBalloon from 'assets/svg/LogoBalloon'
-import Wordmark from 'assets/svg/Logo'
+// assets
+import LogoBalloon from "assets/svg/LogoBalloon"
+import Wordmark from "assets/svg/Logo"
 
-
-const Logo = (props) => {
+// component
+const Logo = props => {
 
   const {
     className,
     type,
-    ...other
+    ...attributes
   } = props
 
-  const classes = classNames(
+  const classes = cx(
     "svg-pink",
     className
   )
@@ -28,7 +28,7 @@ const Logo = (props) => {
   const Tag = type === "balloon" ? LogoBalloon : Wordmark
 
   return (
-    <Tag className={classes} {...other} />
+    <Tag className={classes} {...attributes} />
   )
 }